Remove extra artifacts only for Linux

+++ conanfile.py
@@ -33,8 +33,6 @@ class LibelfConan(ConanFile):
     def configure_autotools(self):
         if not self.autotools:
             args = ['--enable-shared={}'.format('yes' if self.options.shared else 'no')]
-            if self.settings.os == "Macos":
-                args.append('--disable-compat')
             self.autotools = AutoToolsBuildEnvironment(self, win_bash=tools.os_info.is_windows)
             self.autotools.configure(configure_dir=self.source_subfolder, args=args)
         return self.autotools
@@ -49,7 +47,7 @@ class LibelfConan(ConanFile):
         autotools.install()
         shutil.rmtree(os.path.join(self.package_folder, "share"), ignore_errors=True)
         shutil.rmtree(os.path.join(self.package_folder, "lib", "locale"), ignore_errors=True)
-        if self.options.shared:
+        if self.options.shared and self.settings.os == "Linux":
             os.remove(os.path.join(self.package_folder, "lib", "libelf.a"))
 
     def package_info(self):
